[Pallas] [1/3] Move communication primitives from mosaic to core
