+++ jax/_src/pallas/mosaic/core.py
@@ -112,11 +112,6 @@ class TPUMemorySpace(enum.Enum):
     # A convenience function for constructing MemoryRef types.
     return pallas_core.MemoryRef(shape, dtype, self)
 
-class semaphore_dtype(dtypes.extended): pass
-class semaphore(semaphore_dtype): pass
-class dma_semaphore(semaphore_dtype): pass
-class barrier_semaphore(semaphore_dtype): pass
-
 class AbstractSemaphoreTyRules:
   @staticmethod
   def pallas_interpret_element_aval(_) -> jax_core.ShapedArray:
@@ -142,15 +137,15 @@ class AbstractSemaphoreTy(dtypes.ExtendedDType):
 # TODO(sharadmv): implement dtype rules for AbstractSemaphoreTy
 
 class SemaphoreTy(AbstractSemaphoreTy):
-  type = semaphore
+  type = pallas_core.semaphore
   name = "sem"
 
 class DmaSemaphoreTy(AbstractSemaphoreTy):
-  type = dma_semaphore
+  type = pallas_core.dma_semaphore
   name = "dma_sem"
 
 class BarrierSemaphoreTy(AbstractSemaphoreTy):
-  type = barrier_semaphore
+  type = pallas_core.barrier_semaphore
   name = "barrier_sem"
 
 class SemaphoreType(enum.Enum):
